# api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Prompt Injection Defense API
  description: |
    ## Real-time Prompt Injection Detection API
    
    **Enterprise-grade security for AI applications** - Detect and prevent prompt injection attacks with <50ms latency using advanced AI models.
    
    ### Key Features
    
    🛡️ **Multi-Model Detection**
    - ProtectAI DeBERTa v3 (specialized prompt injection classifier)
    - Meta Llama Prompt Guard 2-86M (lightweight jailbreak detection) 
    - Google Gemini 2.0 Flash (advanced semantic analysis)
    
    ⚡ **High Performance**
    - Sub-50ms detection latency
    - Concurrent processing with fallback mechanisms
    - Automatic failover and retry logic
    
    🔐 **Enterprise Security**
    - API key authentication with rate limiting
    - Secure webhook delivery with signature verification
    - Comprehensive audit logging and analytics
    
    ### Threat Types Detected
    
    | Threat Type | Description |
    |-------------|-------------|
    | **Jailbreak** | Attempts to bypass AI safety guardrails |
    | **System Prompt Leak** | Attempts to extract internal AI instructions |
    | **Data Extraction** | Attempts to access training data or user info |
    | **Injection** | Malicious input designed to manipulate AI behavior |
    | **Encoding Attack** | Obfuscated attacks using Base64, hex, etc. |
    | **Delimiter Attack** | Using special characters to break context |
    
    ### Authentication
    
...
    
    ```bash
...
    
...
    ```
    
...
servers:
...
security:
  - APIKeyHeader: []
  - APIKeyBearer: []
...
paths:
...
  /webhooks/register:
    post:
      tags:
        - Webhooks
      summary: Register webhook endpoint
      description: |
        **Webhook Registration** - Register an endpoint to receive detection results and other events.
        
        **Supported Events:**
        - `detection_complete`: Single detection finished
        - `batch_complete`: Batch detection finished
        - `detection_failed`: Detection failed
        - `rate_limit_exceeded`: Rate limit hit
        
        **Security:** Webhooks support secret token verification using HMAC-SHA256.
      operationId: register_webhook
      security:
        - APIKeyHeader: []
        - APIKeyBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
            example:
              url: "https://api.myapp.com/webhooks/prompt-injection"
              events: ["detection_complete", "batch_complete"]
              secret_token: "wh_secret_abc123xyz789"
              description: "Production webhook for prompt injection results"
      responses:
        '200':
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication via X-API-Key header
    APIKeyBearer:
      type: http
      scheme: bearer
      description: API Key authentication via Authorization Bearer header
  schemas:
...
    WebhookRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          maxLength: 500
          description: Webhook endpoint URL
        events:
          type: array
          items:
            type: string
            enum: [detection_complete, batch_complete, detection_failed, rate_limit_exceeded]
          default: ["detection_complete"]
          description: Events to subscribe to
        secret_token:
          type: string
          maxLength: 64
          description: Secret token for webhook verification
        description:
          type: string
          maxLength: 200
          description: Human-readable description
    WebhookResponse:
      type: object
      properties:
        webhook_id:
          type: string
          description: Unique webhook identifier
        url:
          type: string
          format: uri
          description: Webhook endpoint URL
        events:
          type: array
          items:
            type: string
          description: Subscribed events
        secret_token:
          type: string
          description: Secret token (masked)
        description:
          type: string
          description: Description
        is_active:
          type: boolean
          description: Whether webhook is active
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
  responses:
    ValidationError:
      description: Request validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
          example:
            detail:
              - loc: ["body", "text"]
                msg: "field required"
                type: "value_error.missing"
    AuthenticationError:
      description: Authentication error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
          example:
            detail:
              error: "API key required"
              message: "Provide API key in X-API-Key header or Authorization: Bearer header"
...